try again

+++ package.sh
@@ -1,28 +1,22 @@
 #!/usr/bin/env bash
 
 tmp_dir=`date +%Y%m%d%H%M`
-mkdir ${tmp_dir}
-cd ${tmp_dir}
-
-mkdir output src
-
-cd src
-
-wget -q http://download.aqueducts.baidu.com/jdk_1.6.tar.gz && tar xzf jdk_1.6.tar.gz &&  rm -f jdk_1.6.tar.gz
+output_dir=~/nsf/download/logstash/${tmp_dir}
+current_dir=~/nsf/download/logstash/current
 JAVA_HOME=`pwd`/jdk1.6.0_27_x64
 PATH=${JAVA_HOME}/bin:$PATH
 
-#git clone http://gitlab.baidu.com/qudongfang/logstash.git && git checkout aqueducts
-git clone -b aqueducts --single-branch http://gitlab.baidu.com/qudongfang/logstash.git
-cd logstash && make flatjar && cp -f ./build/logstash-*.jar ../output && cd ../
+if [ ! -d "${JAVA_HOME}"]; then  
+  wget -q http://download.aqueducts.baidu.com/jdk_1.6.tar.gz && tar xzf jdk_1.6.tar.gz &&  rm -f jdk_1.6.tar.gz
+fi  
 
-#git clone http://gitlab.baidu.com/qudongfang/logstash-kafka.git && git checkout aqueducts
-git clone -b aqueducts --single-branch http://gitlab.baidu.com/qudongfang/logstash-kafka.git
-cd logstash-kafka && make flatjar && cp -f ./build/logstash-*.jar ../output && cd ../
+mkdir -p ${tmp_dir} ${output_dir}
+cd ${tmp_dir}
 
-cd ../output
+git clone -b aqueducts http://gitlab.baidu.com/qudongfang/logstash.git
+cd logstash && make flatjar && cp -f ./build/logstash-*.jar ${output_dir} && cd ../
 
-mkdir -p ~/nfs/download/${tmp_dir}/
-cp -f logstash-kafka-*.jar ~/nfs/download/${tmp_dir}/ 
-
+git clone -b aqueducts http://gitlab.baidu.com/qudongfang/logstash-kafka.git
+cd logstash-kafka && make flatjar
+mkdir -p ~/nfs/download/logstash/${tmp_dir}/ && cp -f ./build/logstash-*.jar ${outout_dir} && rm -f ${current_dir} && ln -s ${output_dir} ${current_dir}
